requirements.txt
streamlit==1.48.0
langchain==0.3.27
langchain-community==0.3.27
langchain-ollama==0.3.6
PyPDF2==3.0.1
pytesseract==0.3.13
Pillow==10.4.0
requests==2.32.4
faiss-cpu==1.11.0.post1
ddgs==9.5.2
python-dotenv==1.1.1
